judyp: Patch com.qti.chi.override.so to fix viewfinder issues
Taken from stock firmware (V350AWM30c) with the main purpose of
eliminating as much as LG bloated stuff in the library as much as
possible, only keeping the needed one on AOSP.

This library also contains buffer fix for the device, which solve
the viewfinder lag issue (it displays everything at 15fps) in camera
apps. (e.g. Google Camera)

Thanks to Wyroczen for the patch.

Change-Id: I0a8de1362b7b8d5f8d4251bd63beb9b23ea88f57